Feature/bmspt 177 bcf to stream (#10)
* [BMSPT-177] implementation of bcf to stream functionality

* [BMSPT-177] unit tests for stream handling, stream reading fix

* Renaming input arguments from targetFolder to target

+++ src/bcf-toolkit/Worker/Bcf21/Converter.cs
@@ -126,6 +126,28 @@ public class Converter : IConverter {
   }
 
   /// <summary>
+  ///   The method handles the BCF content from the given objects to the
+  ///   specified stream.
+  /// </summary>
+  /// <param name="bcf">The `IBcf` interface of the BCF.</param>
+  /// <returns></returns>
+  /// <exception cref="FileNotFoundException"></exception>
+  public async Task<Stream> BcfStream(IBcf bcf) {
+    var workingDir = Directory.GetCurrentDirectory();
+    var bcfTargetPath = workingDir + "/bcf.bcfzip";
+
+    var tmpFolder = await WriteBcf(bcfTargetPath, (Bcf)bcf, false);
+
+    var stream = new FileStream(bcfTargetPath, FileMode.Open, FileAccess.Read);
+
+    // After the filestream is ready we can delete the folders
+    Directory.Delete(tmpFolder, true);
+    File.Delete(bcfTargetPath);
+
+    return stream;
+  }
+
+  /// <summary>
   ///   The method writes the BCF content from the given objects to the
   ///   specified target and compresses it.
   ///   The markups will be written into the topic folder structure:
@@ -138,9 +160,10 @@ public class Converter : IConverter {
   /// </summary>
   /// <param name="target">The target file name of the BCFzip.</param>
   /// <param name="bcf">The BCF object.</param>
-  /// <returns></returns>
+  /// <param name="delete">Should delete the generated tmp folder now or later</param>
+  /// <returns>Generated temp folder path</returns>
   /// <exception cref="ApplicationException"></exception>
-  private static async Task WriteBcf(string target, Bcf bcf) {
+  private static async Task<string> WriteBcf(string target, Bcf bcf, bool delete = true) {
     var targetFolder = Path.GetDirectoryName(target);
     if (targetFolder == null)
       throw new ApplicationException(
@@ -148,7 +171,9 @@ public class Converter : IConverter {
 
     // Will create a tmp folder for the intermediate files.
     var tmpFolder = $"{targetFolder}/tmp";
+
     if (Directory.Exists(tmpFolder)) Directory.Delete(tmpFolder, true);
+
     Directory.CreateDirectory(tmpFolder);
 
     var tasks = new List<Task>();
@@ -199,7 +224,11 @@ public class Converter : IConverter {
     Console.WriteLine($"Zipping the output: {target}");
     if (File.Exists(target)) File.Delete(target);
     ZipFile.CreateFromDirectory(tmpFolder, target);
-    Directory.Delete(tmpFolder, true);
+
+    if (delete)
+      Directory.Delete(tmpFolder, true);
+
+    return tmpFolder;
   }
 
   /// <summary>
